# api/openapi/cbizdocsmanager-v1.yaml
openapi: 3.0.3
info:
  title: CBizDocsManager API
  version: 1.0.0
  description: 公务单据管理系统后端 API v1 契约。
servers:
  - url: /
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/platform/groups:
    post:
      tags: [Platform]
      operationId: createGroup
      summary: 平台管理员创建组及主账号
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateGroupRequest'
      responses:
        '201':
          description: 组与主账号创建成功
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateGroupApiResponse'
        '400':
          description: 请求参数错误（VALIDATION_FAILED）
          x-error-codes: [VALIDATION_FAILED]
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorApiResponse'
        '401':
          $ref: '#/components/responses/TokenExpired'
        '403':
          description: 无平台管理权限或必须先修改初始密码（FORBIDDEN、AUTH_PASSWORD_CHANGE_REQUIRED）
          x-error-codes: [FORBIDDEN, AUTH_PASSWORD_CHANGE_REQUIRED]
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorApiResponse'
        '409':
          description: 组名或主账号用户名已存在（GROUP_NAME_EXISTS、USER_USERNAME_EXISTS）
          x-error-codes: [GROUP_NAME_EXISTS, USER_USERNAME_EXISTS]
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorApiResponse'
        '500':
          $ref: '#/components/responses/InternalError'
    get:
      tags: [Platform]
      operationId: listGroups
      summary: 平台管理员分页查询业务组
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: page_size
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: status
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/GroupStatus'
        - name: keyword
          in: query
          required: false
          description: 按组名模糊匹配
          schema:
            type: string
      responses:
        '200':
          description: 查询成功
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GroupPageApiResponse'
        '400':
          description: 请求参数错误（VALIDATION_FAILED）
          x-error-codes: [VALIDATION_FAILED]
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorApiResponse'
        '401':
          $ref: '#/components/responses/TokenExpired'
        '403':
          description: 无平台管理权限或必须先修改初始密码（FORBIDDEN、AUTH_PASSWORD_CHANGE_REQUIRED）
          x-error-codes: [FORBIDDEN, AUTH_PASSWORD_CHANGE_REQUIRED]
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorApiResponse'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/platform/groups/{group_id}:
    get:
      tags: [Platform]
      operationId: getGroupDetail
      summary: 平台管理员查询组治理详情
      parameters:
        - $ref: '#/components/parameters/GroupIdPath'
      responses:
        '200':
          description: 查询成功
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GroupDetailApiResponse'
        '400':
          description: 请求参数错误（VALIDATION_FAILED）
          x-error-codes: [VALIDATION_FAILED]
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorApiResponse'
        '401':
          $ref: '#/components/responses/TokenExpired'
        '403':
          description: 无平台管理权限或必须先修改初始密码（FORBIDDEN、AUTH_PASSWORD_CHANGE_REQUIRED）
          x-error-codes: [FORBIDDEN, AUTH_PASSWORD_CHANGE_REQUIRED]
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorApiResponse'
        '404':
          description: 组不存在（GROUP_NOT_FOUND）
          x-error-codes: [GROUP_NOT_FOUND]
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorApiResponse'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/platform/groups/{group_id}/status:
    patch:
      tags: [Platform]
      operationId: changeGroupStatus
      summary: 平台管理员停用或启用业务组
      description: 目标状态与当前状态一致时按幂等成功处理；停用组会连同撤销该组全部刷新会话。
      parameters:
        - $ref: '#/components/parameters/GroupIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangeGroupStatusRequest'
      responses:
        '200':
          description: 变更成功，返回组最新摘要
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GroupSummaryApiResponse'
        '400':
          description: 请求参数错误（VALIDATION_FAILED）
          x-error-codes: [VALIDATION_FAILED]
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorApiResponse'
        '401':
          $ref: '#/components/responses/TokenExpired'
        '403':
          description: 无平台管理权限或必须先修改初始密码（FORBIDDEN、AUTH_PASSWORD_CHANGE_REQUIRED）
          x-error-codes: [FORBIDDEN, AUTH_PASSWORD_CHANGE_REQUIRED]
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorApiResponse'
        '404':
          description: 组不存在（GROUP_NOT_FOUND）
          x-error-codes: [GROUP_NOT_FOUND]
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorApiResponse'
        '409':
          description: 组已被其他请求修改（RESOURCE_VERSION_CONFLICT）
          x-error-codes: [RESOURCE_VERSION_CONFLICT]
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorApiResponse'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/platform/groups/{group_id}/owner:
    put:
      tags: [Platform]
      operationId: changeGroupOwner
      summary: 平台管理员交接组主账号
      description: 支持把现有成员提升为主账号（existing_member）或新建账号并设为主账号（new_account）。
      parameters:
        - $ref: '#/components/parameters/GroupIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangeGroupOwnerRequest'
      responses:
        '200':
          description: 交接成功
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OwnerChangedApiResponse'
        '400':
          description: 请求参数错误或交接目标无效（VALIDATION_FAILED、OWNER_TARGET_INVALID）
          x-error-codes: [VALIDATION_FAILED, OWNER_TARGET_INVALID]
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorApiResponse'
        '401':
          $ref: '#/components/responses/TokenExpired'
        '403':
          description: 无平台管理权限或必须先修改初始密码（FORBIDDEN、AUTH_PASSWORD_CHANGE_REQUIRED）
          x-error-codes: [FORBIDDEN, AUTH_PASSWORD_CHANGE_REQUIRED]
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorApiResponse'
        '404':
          description: 组不存在（GROUP_NOT_FOUND）
          x-error-codes: [GROUP_NOT_FOUND]
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorApiResponse'
        '409':
          description: 组已被其他请求修改或用户名已存在（RESOURCE_VERSION_CONFLICT、USER_USERNAME_EXISTS）
          x-error-codes: [RESOURCE_VERSION_CONFLICT, USER_USERNAME_EXISTS]
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorApiResponse'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  headers:
    RequestID:
      description: 本次请求的唯一追踪 ID
      schema:
        type: string
  parameters:
    GroupIdPath:
      name: group_id
      in: path
      required: true
      schema:
        type: integer
        format: int64
        minimum: 1
...
  responses:
    TokenExpired:
      description: 访问令牌无效或已过期（AUTH_TOKEN_EXPIRED）
      x-error-codes: [AUTH_TOKEN_EXPIRED]
      headers:
        X-Request-ID:
          $ref: '#/components/headers/RequestID'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorApiResponse'
    InternalError:
      description: 服务内部错误（INTERNAL_ERROR）
      x-error-codes: [INTERNAL_ERROR]
      headers:
        X-Request-ID:
          $ref: '#/components/headers/RequestID'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorApiResponse'
  schemas:
    ErrorCode:
      type: string
      enum:
        - VALIDATION_FAILED
        - AUTH_INVALID_CREDENTIALS
        - AUTH_TOKEN_EXPIRED
        - AUTH_REFRESH_INVALID
        - AUTH_PASSWORD_CHANGE_REQUIRED
        - USER_USERNAME_EXISTS
        - INVITATION_INVALID
        - INVITATION_EXPIRED
        - INVITATION_USED
        - INVITATION_NOT_FOUND
        - INVITATION_NOT_REVEALABLE
        - INVITATION_NOT_REVOKABLE
        - INVITATION_DECRYPT_FAILED
        - GROUP_NAME_EXISTS
        - GROUP_NOT_FOUND
        - OWNER_TARGET_INVALID
        - MEMBER_NOT_FOUND
        - MEMBER_OWNER_PROTECTED
        - MEMBER_SELF_OPERATION_FORBIDDEN
        - PERMISSION_CODE_INVALID
        - DICTIONARY_NOT_FOUND
        - DICTIONARY_NAME_EXISTS
        - DICTIONARY_PARENT_INVALID
        - RESOURCE_VERSION_CONFLICT
        - CSRF_INVALID
        - ORIGIN_FORBIDDEN
        - FORBIDDEN
        - INTERNAL_ERROR
    FieldError:
      type: object
      additionalProperties: false
      required: [field, message]
      properties:
        field:
          type: string
        message:
          type: string
    ApiResponse:
      type: object
      required: [code, message, data, request_id]
      properties:
        code:
          type: string
          description: 成功标识或稳定错误码
        message:
          type: string
        data:
          type: object
          nullable: true
          additionalProperties: true
        request_id:
          type: string
        field_errors:
          type: array
          items:
            $ref: '#/components/schemas/FieldError'
    ErrorApiResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          required: [code, message, data, request_id]
          properties:
            code:
              $ref: '#/components/schemas/ErrorCode'
            data:
              type: object
              nullable: true
            field_errors:
              type: array
              items:
                $ref: '#/components/schemas/FieldError'
...
    CreateGroupRequest:
      type: object
      additionalProperties: false
      required: [name, owner_username, owner_display_name, owner_temporary_password]
      properties:
        name:
          type: string
          minLength: 1
        owner_username:
          type: string
          minLength: 1
        owner_display_name:
          type: string
          minLength: 1
        owner_temporary_password:
          type: string
          format: password
          minLength: 8
...
    UserSummary:
      type: object
      additionalProperties: false
      required: [id, username, display_name, account_type]
      properties:
        id:
          type: integer
          format: int64
          minimum: 1
        username:
          type: string
        display_name:
          type: string
        account_type:
          type: string
          enum: [platform_admin, group_owner, member]
    GroupSummary:
      type: object
      additionalProperties: false
      required: [id, name]
      properties:
        id:
          type: integer
          format: int64
          minimum: 1
        name:
          type: string
...
    GroupCreatedData:
      type: object
      additionalProperties: false
      required: [group, owner]
      properties:
        group:
          $ref: '#/components/schemas/GroupSummary'
        owner:
          $ref: '#/components/schemas/UserSummary'
    CreateGroupApiResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/GroupCreatedData'
...
    GroupStatus:
      type: string
      enum: [active, disabled]
    GroupSummaryData:
      type: object
      additionalProperties: false
      required: [id, name, status, owner, member_count, version, created_at, updated_at]
      properties:
        id:
          type: integer
          format: int64
          minimum: 1
        name:
          type: string
        status:
          $ref: '#/components/schemas/GroupStatus'
        owner:
          $ref: '#/components/schemas/UserSummary'
        member_count:
          type: integer
          format: int64
          minimum: 0
        version:
          type: integer
          format: int64
          minimum: 1
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    GroupPageData:
      type: object
      additionalProperties: false
      required: [items, page, page_size, total]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/GroupSummaryData'
        page:
          type: integer
          minimum: 1
        page_size:
          type: integer
          minimum: 1
        total:
          type: integer
          format: int64
          minimum: 0
    GroupPageApiResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/GroupPageData'
    GroupSummaryApiResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/GroupSummaryData'
    OwnerCandidateData:
      type: object
      additionalProperties: false
      required: [membership_id, user]
      properties:
        membership_id:
          type: integer
          format: int64
          minimum: 1
        user:
          $ref: '#/components/schemas/UserSummary'
    GroupDetailData:
      type: object
      additionalProperties: false
      required: [group, member_counts, owner_candidates]
      properties:
        group:
          $ref: '#/components/schemas/GroupSummaryData'
        member_counts:
          type: object
          description: 按成员状态聚合的人数，键为 active / disabled / removed。
          additionalProperties:
            type: integer
            format: int64
            minimum: 0
        owner_candidates:
          type: array
          description: 可作为新主账号的现有成员（仅 active 成员）。
          items:
            $ref: '#/components/schemas/OwnerCandidateData'
    GroupDetailApiResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/GroupDetailData'
    ChangeGroupStatusRequest:
      type: object
      additionalProperties: false
      required: [status, version]
      properties:
        status:
          $ref: '#/components/schemas/GroupStatus'
        version:
          type: integer
          format: int64
          minimum: 1
          description: 客户端持有的组版本号，用于乐观锁校验。
    ChangeGroupOwnerRequest:
      type: object
      additionalProperties: false
      required: [mode, version]
      properties:
        mode:
          type: string
          enum: [existing_member, new_account]
        membership_id:
          type: integer
          format: int64
          minimum: 1
          nullable: true
          description: mode=existing_member 时必填，指向候选成员的关系 ID。
        username:
          type: string
          nullable: true
          description: mode=new_account 时必填。
        display_name:
          type: string
          nullable: true
          description: mode=new_account 时必填。
        temporary_password:
          type: string
          format: password
          nullable: true
          minLength: 8
          description: mode=new_account 时必填，长度不少于 8。
        version:
          type: integer
          format: int64
          minimum: 1
    OwnerChangedData:
      type: object
      additionalProperties: false
      required: [group, owner]
      properties:
        group:
          $ref: '#/components/schemas/GroupSummaryData'
        owner:
          $ref: '#/components/schemas/UserSummary'
    OwnerChangedApiResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/OwnerChangedData'
...